src/data/sentiment_analyzer.py
"""
Module for analyzing sentiment from news and social media sources related to Bitcoin trading.
"""
import logging
from typing import Dict, List, Optional, Tuple
import pandas as pd
import numpy as np
from datetime import datetime, timedelta

logger = logging.getLogger(__name__)

class SentimentAnalyzer:
    """Class for fetching and analyzing sentiment data from various sources."""
    
    REQUIRED_CONFIG_KEYS = {'news', 'social', 'analysis'}
    
    def __init__(self, config: Optional[Dict] = None):
        """
        Initialize the sentiment analyzer.
        
        Args:
            config: Configuration dictionary containing API keys and settings
            
        Raises:
            ValueError: If config is provided but missing required keys
        """
        self.config = config or {}
        
        if self.config and not all(key in self.config for key in self.REQUIRED_CONFIG_KEYS):
            raise ValueError(
                f"Config must contain the following keys: {self.REQUIRED_CONFIG_KEYS}"
            )
        
        self._initialize_apis()
        
    def _initialize_apis(self):
        """Initialize connections to various news and social media APIs."""
        # TODO: Implement API initialization for:
        # - News APIs (e.g., NewsAPI, CryptoCompare News)
        # - Twitter API
        # - Reddit API
        pass
        
    async def fetch_news_sentiment(self, 
                                 start_time: datetime,
                                 end_time: datetime) -> pd.DataFrame:
        """
        Fetch and analyze sentiment from news articles.
        
        Args:
            start_time: Start time for fetching news
            end_time: End time for fetching news
            
        Returns:
            DataFrame containing news sentiment scores
        """
        # TODO: Implement news fetching and sentiment analysis
        # For testing, return sample data
        dates = pd.date_range(start=start_time, end=end_time, freq='1h')
        sentiment_data = {
            'timestamp': dates,
            'source': ['NewsAPI'] * len(dates),
            'title': [f'Test News {i}' for i in range(len(dates))],
            'sentiment_score': np.random.uniform(-1, 1, len(dates)),
            'impact_score': np.random.uniform(0, 1, len(dates))
        }
        return pd.DataFrame(sentiment_data)
        
    async def fetch_social_sentiment(self,
                                   start_time: datetime,
                                   end_time: datetime) -> pd.DataFrame:
        """
        Fetch and analyze sentiment from social media.
        
        Args:
            start_time: Start time for fetching social media data
            end_time: End time for fetching social media data
            
        Returns:
            DataFrame containing social media sentiment scores
        """
        # TODO: Implement social media sentiment analysis
        # For testing, return sample data
        dates = pd.date_range(start=start_time, end=end_time, freq='1h')
        sentiment_data = {
            'timestamp': dates,
            'platform': ['Twitter'] * len(dates),
            'content_type': ['tweet'] * len(dates),
            'sentiment_score': np.random.uniform(-1, 1, len(dates)),
            'engagement_score': np.random.uniform(0, 1000, len(dates))
        }
        return pd.DataFrame(sentiment_data)
        
    def aggregate_sentiment(self,
                          news_sentiment: pd.DataFrame,
                          social_sentiment: pd.DataFrame,
                          window: str = '1h') -> pd.DataFrame:
        """
        Aggregate sentiment scores from different sources.
        
        Args:
            news_sentiment: DataFrame containing news sentiment
            social_sentiment: DataFrame containing social sentiment
            window: Resampling window for aggregation
            
        Returns:
            DataFrame containing aggregated sentiment scores
        """
        # Ensure timestamps are datetime and set as index
        news_sentiment = news_sentiment.copy()
        social_sentiment = social_sentiment.copy()
        
        news_sentiment['timestamp'] = pd.to_datetime(news_sentiment['timestamp'])
        social_sentiment['timestamp'] = pd.to_datetime(social_sentiment['timestamp'])
        
        news_sentiment.set_index('timestamp', inplace=True)
        social_sentiment.set_index('timestamp', inplace=True)
        
        # Calculate weighted sentiment scores
        news_weights = news_sentiment['impact_score']
        social_weights = social_sentiment['engagement_score'] / social_sentiment['engagement_score'].max()
        
        news_weighted = (news_sentiment['sentiment_score'] * news_weights).resample(window).mean()
        social_weighted = (social_sentiment['sentiment_score'] * social_weights).resample(window).mean()
        
        # Combine sentiment scores
        sentiment_df = pd.DataFrame({
            'news_sentiment': news_weighted,
            'social_sentiment': social_weighted
        })
        
        # Calculate composite sentiment and confidence
        sentiment_df['composite_sentiment'] = (
            0.6 * sentiment_df['news_sentiment'] +
            0.4 * sentiment_df['social_sentiment']
        )
        
        sentiment_df['confidence_score'] = 1 - np.abs(
            sentiment_df['news_sentiment'] - sentiment_df['social_sentiment']
        ) / 2
        
        # Reset index to keep timestamp as column
        sentiment_df.reset_index(inplace=True)
        
        return sentiment_df
        
    def calculate_sentiment_features(self,
                                   sentiment_df: pd.DataFrame) -> pd.DataFrame:
        """
        Calculate additional features from sentiment data.
        
        Args:
            sentiment_df: DataFrame containing sentiment data
            
        Returns:
            DataFrame with additional sentiment features
        """
        features = sentiment_df.copy()
        
        # Ensure timestamp is index for calculations
        features.set_index('timestamp', inplace=True)
        
        # Calculate sentiment momentum (change over time)
        features['sentiment_momentum'] = features['composite_sentiment'].diff()
        
        # Calculate sentiment volatility
        features['sentiment_volatility'] = features['composite_sentiment'].rolling(
            window=12, min_periods=1
        ).std()
        
        # Identify extreme sentiment
        sentiment_std = features['composite_sentiment'].std()
        features['extreme_sentiment'] = (np.abs(features['composite_sentiment']) > 
                                       (2 * sentiment_std)).astype(float)
        
        # Calculate sentiment divergence (difference between news and social)
        features['sentiment_divergence'] = np.abs(
            features['news_sentiment'] - features['social_sentiment']
        )
        
        # Reset index to keep timestamp as column
        features.reset_index(inplace=True)
        
        return features
